--- a/pkg/usagetracker/tenantshard/event.go
+++ b/pkg/usagetracker/tenantshard/event.go
@@ -8,50 +8,107 @@ import (
 	"github.com/grafana/mimir/pkg/usagetracker/clock"
 )
 
-type EventType uint8
+func Track(refs []uint64, value clock.Minutes, series *atomic.Uint64, limit uint64, resp chan TrackResponse) Event {
+	return Event{
+		typ:           track,
+		series:        series,
+		limit:         limit,
+		refs:          refs,
+		timestamp:     value,
+		trackResponse: resp,
+	}
+}
+
+func Create(refs []uint64, ts clock.Minutes, series *atomic.Uint64, done chan struct{}) Event {
+	return Event{
+		typ:       create,
+		series:    series,
+		refs:      refs,
+		timestamp: ts,
+		done:      done,
+	}
+}
+
+func Load(refs []RefTimestamp, series *atomic.Uint64, done chan struct{}) Event {
+	return Event{
+		typ:      load,
+		series:   series,
+		loadRefs: refs,
+		done:     done,
+	}
+}
+
+func Clone(cloner chan<- func(LengthCallback, IteratorCallback)) Event {
+	return Event{
+		typ:    clone,
+		cloner: cloner,
+	}
+}
+
+func Cleanup(watermark clock.Minutes, series *atomic.Uint64, done chan struct{}) Event {
+	return Event{
+		typ:       cleanup,
+		series:    series,
+		watermark: watermark,
+		done:      done,
+	}
+}
+
+func Shutdown() Event {
+	return Event{
+		typ: shutdown,
+	}
+}
+
+type RefTimestamp struct {
+	Ref       uint64
+	Timestamp clock.Minutes
+}
+
+type eventType uint8
 
 const (
-	TrackSeries EventType = iota
-	Load
-	Clone
-	Cleanup
-	Shutdown
+	track eventType = iota
+	create
+	load
+	clone
+	cleanup
+	shutdown
 )
 
 type Event struct {
-	Type EventType
+	typ eventType
 
-	// Series is sent in all events.
-	Series *atomic.Uint64
+	// Series is sent Track, Load and Cleanup.
+	series *atomic.Uint64
 
-	// Limit is sent in TrackSeries event, it could be math.MaxUint64.
-	Limit uint64
+	// Limit is sent in Track event, it could be math.MaxUint64.
+	limit uint64
 
-	// Refs is sent in TrackSeries event.
-	Refs []uint64
+	// Refs is sent in Track and Create events.
+	refs []uint64
 
-	// Value is sent in TrackSeries event.
-	Value clock.Minutes
+	// Timestamp is sent in Track and Create events.
+	timestamp clock.Minutes
 
 	// LoadRefs is sent in Load event.
-	// TODO: we could optimize allocations making this a slice of ref/ts pairs.
-	LoadRefs map[uint64]clock.Minutes
+	loadRefs []RefTimestamp
 
 	// Watermark is sent in Cleanup event.
-	Watermark clock.Minutes
+	watermark clock.Minutes
 
-	// TrackSeriesResponse is sent in TrackSeries event.
-	TrackSeriesResponse chan TrackSeriesResponse
+	// TrackResponse is sent in Track event.
+	trackResponse chan TrackResponse
 
 	// Cloner is sent in Clone event.
 	// The received function will call LengthCallback first, and then IteratorCallback for each one of the series.
-	Cloner chan<- func(LengthCallback, IteratorCallback)
+	cloner chan<- func(LengthCallback, IteratorCallback)
 
 	// Done is sent in Cleanup event.
-	Done chan struct{}
+	done chan struct{}
 }
 
-type TrackSeriesResponse struct {
+type TrackResponse struct {
 	Created  []uint64
 	Rejected []uint64
 }
@@ -61,36 +118,42 @@ func (m *Map) Events() chan<- Event { return m.events }
 
 func (m *Map) worker() {
 	for ev := range m.events {
-		switch ev.Type {
-		case TrackSeries:
-			var resp TrackSeriesResponse
-			for _, ref := range ev.Refs {
-				if created, rejected := m.put(ref, ev.Value, ev.Series, ev.Limit, true); created {
+		switch ev.typ {
+		case track:
+			var resp TrackResponse
+			for _, ref := range ev.refs {
+				if created, rejected := m.put(ref, ev.timestamp, ev.series, ev.limit, true); created {
 					resp.Created = append(resp.Created, ref)
 				} else if rejected {
 					resp.Rejected = append(resp.Rejected, ref)
 				}
 			}
-			ev.TrackSeriesResponse <- resp
+			ev.trackResponse <- resp
 			// TODO: we could use this idle time to resize even larger.
 
-		case Load:
-			for ref, v := range ev.LoadRefs {
-				m.put(ref, v, ev.Series, 0, false)
+		case create:
+			for _, ref := range ev.refs {
+				m.put(ref, ev.timestamp, ev.series, 0, false)
 			}
-			ev.Done <- struct{}{}
+			ev.done <- struct{}{}
 
-		case Clone:
-			ev.Cloner <- m.cloner()
+		case load:
+			for _, entry := range ev.loadRefs {
+				m.put(entry.Ref, entry.Timestamp, ev.series, 0, false)
+			}
+			ev.done <- struct{}{}
 
-		case Cleanup:
-			m.cleanup(ev.Watermark, ev.Series)
+		case clone:
+			ev.cloner <- m.cloner()
+
+		case cleanup:
+			m.cleanup(ev.watermark, ev.series)
 			if m.dead > m.limit/2 {
 				m.rehash(m.nextSize())
 			}
-			ev.Done <- struct{}{}
+			ev.done <- struct{}{}
 
-		case Shutdown:
+		case shutdown:
 			return
 		}
 	}