… cmd/mimir/config-descriptor.json
… docs/sources/mimir/configure/configuration-parameters/index.md
… pkg/usagetracker/tenantshard/map_test.go
… pkg/usagetracker/tracker_store_bench_test.go
… pkg/usagetracker/tracker_store_collector.go
… pkg/usagetracker/tracker_store_snapshot.go
… pkg/usagetracker/tracker_store_test.go