This refactors the entire tracker.

There are two big changes here: how the map is implemented and how we interact with it.

How the map is implemented:

We're getting rid of usual maps storing atomic.Int32: it's an overkill to have 32bit int to store a byte, it's even more an overkill to store a 64bit pointer to that, let's not talk about map overhead.

Instead of that, we're building kind of swiss map, I based the implementation on the code from dolthub/swiss, you can read a blog post how that works here:
https://www.dolthub.com/blog/2023-03-28-swiss-map/

However, most of the code was replaced.

What did I do? Swiss map has a slice groups of hash suffixes, which are one byte, to quickly find the bucket where our element is, and it also has a slice of groups of keys. That's still a lot of overhead for our use.

What are our restrictions:
- Our keys are uint64, so great, we don't need to hash, our keys are hashes already!.
- All our keys in a single shard have the same suffix of 7 bytes (because we have 128 shards).

So, I changed the implementation to "index" based on the prefix, not the suffix of the key (series hash), and since the suffix is always the same, we could use that space for something else... Guess what? Our values are also 7 bits, so we're storing the key AND the value in a single uint64 field. Now each series is just uint64+uint8 in the map implementation (plus the overallocation of the slice, of course).

I didn't just store the value in the suffix of the uint64 data, but I also negated it, so we can figure out whether a value is a real value or an empty value without having to check if there's a tombstone in the index! Cleanup is now just iterating one single slice, isn't that awesome?

Well, but how do we coordinate all of that? We can't do atomic byte operations (can we? IDK), but also we probably don't want to. I guessed (to be investigated whether that was true) that lots of CPU usage comes from having to do atomic-everything.

I head that golang is good at handling hundreds of goroutines (and I managed tens of thousands of them in the past, although not in a high-performance env), so I decided to just run one goroutine per each map shard.

Each map shard has its own worker which takes care of updating the internal state: now the synchronization overhad is in pushing the data to the workers through a channel. But once data is in the worker, a single shard is always updated by the same goroutine.

Say we have 10k tenants in an env, that's 10K goroutines, not that bad... I hope. We'll see.

I also considered lowering the number of shards to 64, ~but in order to achieve that we need to restrict the idleTimeout to 30 minutes, so we can fit the minutes timestamp in a 64 byte value. I think we can do that.~ We can fit any number of shards if we shift the ref stored in data by 7 bits to the left, as we already have those bits in the index anyway.

Benchmarks look great, the improvements get better as the load grows, especially when cleanup p goroutine is running:
```
goos: darwin
goarch: arm64
pkg: github.com/grafana/mimir/pkg/usagetracker
cpu: Apple M3 Pro
                                                                                             │ map_atomic_ptrs │              swiss_chans              │
                                                                                             │     sec/op      │    sec/op      vs base                │
TrackerStoreTrackSeries/totalSeries=10M/seriesPerRequest=1K/tenants=1/cleanups=true-12            71.68µ ± 12%   138.49µ ±  6%   +93.21% (p=0.002 n=6)
TrackerStoreTrackSeries/totalSeries=10M/seriesPerRequest=1K/tenants=1/cleanups=false-12           16.35µ ±  1%    28.77µ ±  1%   +75.91% (p=0.002 n=6)
TrackerStoreTrackSeries/totalSeries=10M/seriesPerRequest=1K/tenants=10/cleanups=true-12           31.69µ ±  5%    69.19µ ±  2%  +118.37% (p=0.002 n=6)
TrackerStoreTrackSeries/totalSeries=10M/seriesPerRequest=1K/tenants=10/cleanups=false-12          16.15µ ±  3%    23.59µ ±  1%   +46.07% (p=0.002 n=6)
TrackerStoreTrackSeries/totalSeries=10M/seriesPerRequest=1K/tenants=100/cleanups=true-12          46.96µ ± 30%    34.10µ ±  4%         ~ (p=0.240 n=6)
TrackerStoreTrackSeries/totalSeries=10M/seriesPerRequest=1K/tenants=100/cleanups=false-12         16.39µ ±  2%    24.74µ ±  1%   +50.94% (p=0.002 n=6)
TrackerStoreTrackSeries/totalSeries=10M/seriesPerRequest=1K/tenants=1000/cleanups=true-12      12480.17µ ±  6%    38.47µ ± 18%   -99.69% (p=0.002 n=6)
TrackerStoreTrackSeries/totalSeries=10M/seriesPerRequest=1K/tenants=1000/cleanups=false-12        11.98µ ±  1%    25.52µ ±  1%  +112.99% (p=0.002 n=6)
TrackerStoreTrackSeries/totalSeries=10M/seriesPerRequest=10K/tenants=1/cleanups=true-12           211.0µ ± 10%    354.9µ ±  2%   +68.24% (p=0.002 n=6)
TrackerStoreTrackSeries/totalSeries=10M/seriesPerRequest=10K/tenants=1/cleanups=false-12          137.3µ ±  1%    110.6µ ±  2%   -19.42% (p=0.002 n=6)
TrackerStoreTrackSeries/totalSeries=10M/seriesPerRequest=10K/tenants=10/cleanups=true-12          208.6µ ± 20%    175.5µ ±  3%   -15.89% (p=0.009 n=6)
TrackerStoreTrackSeries/totalSeries=10M/seriesPerRequest=10K/tenants=10/cleanups=false-12         132.8µ ±  2%    101.6µ ±  1%   -23.53% (p=0.002 n=6)
TrackerStoreTrackSeries/totalSeries=10M/seriesPerRequest=10K/tenants=100/cleanups=true-12        161.90µ ± 16%    94.78µ ±  3%   -41.46% (p=0.002 n=6)
TrackerStoreTrackSeries/totalSeries=10M/seriesPerRequest=10K/tenants=100/cleanups=false-12        90.56µ ±  2%    77.28µ ±  1%   -14.67% (p=0.002 n=6)
TrackerStoreTrackSeries/totalSeries=100M/seriesPerRequest=1K/tenants=1/cleanups=true-12          11.997m ±  7%    1.382m ±  9%   -88.48% (p=0.002 n=6)
TrackerStoreTrackSeries/totalSeries=100M/seriesPerRequest=1K/tenants=1/cleanups=false-12          18.77µ ±  5%    30.23µ ± 14%   +61.07% (p=0.002 n=6)
TrackerStoreTrackSeries/totalSeries=100M/seriesPerRequest=1K/tenants=10/cleanups=true-12         11.142m ± 11%    1.489m ± 10%   -86.64% (p=0.002 n=6)
TrackerStoreTrackSeries/totalSeries=100M/seriesPerRequest=1K/tenants=10/cleanups=false-12         17.96µ ±  3%    25.53µ ±  2%   +42.11% (p=0.002 n=6)
TrackerStoreTrackSeries/totalSeries=100M/seriesPerRequest=1K/tenants=100/cleanups=true-12       11520.7µ ± 48%    167.9µ ±  4%   -98.54% (p=0.002 n=6)
TrackerStoreTrackSeries/totalSeries=100M/seriesPerRequest=1K/tenants=100/cleanups=false-12        18.27µ ± 66%    27.77µ ±  1%         ~ (p=0.065 n=6)
TrackerStoreTrackSeries/totalSeries=100M/seriesPerRequest=1K/tenants=1000/cleanups=true-12      20411.1µ ±  3%    191.5µ ± 72%   -99.06% (p=0.002 n=6)
TrackerStoreTrackSeries/totalSeries=100M/seriesPerRequest=1K/tenants=1000/cleanups=false-12       18.78µ ± 24%    26.78µ ±  6%   +42.64% (p=0.002 n=6)
TrackerStoreTrackSeries/totalSeries=100M/seriesPerRequest=10K/tenants=1/cleanups=true-12         13.098m ± 44%    2.635m ± 21%   -79.88% (p=0.002 n=6)
TrackerStoreTrackSeries/totalSeries=100M/seriesPerRequest=10K/tenants=1/cleanups=false-12         153.3µ ±  9%    130.0µ ±  7%   -15.15% (p=0.002 n=6)
TrackerStoreTrackSeries/totalSeries=100M/seriesPerRequest=10K/tenants=10/cleanups=true-12        11.623m ± 12%    1.750m ± 10%   -84.94% (p=0.002 n=6)
TrackerStoreTrackSeries/totalSeries=100M/seriesPerRequest=10K/tenants=10/cleanups=false-12        156.3µ ± 25%    119.4µ ±  1%   -23.59% (p=0.002 n=6)
TrackerStoreTrackSeries/totalSeries=100M/seriesPerRequest=10K/tenants=100/cleanups=true-12       10.716m ± 18%    1.234m ±  4%   -88.49% (p=0.002 n=6)
TrackerStoreTrackSeries/totalSeries=100M/seriesPerRequest=10K/tenants=100/cleanups=false-12       140.9µ ±  7%    110.1µ ± 19%   -21.84% (p=0.002 n=6)
TrackerStoreTrackSeries/totalSeries=100M/seriesPerRequest=10K/tenants=1000/cleanups=true-12     20809.0µ ±  4%    275.3µ ±  2%   -98.68% (p=0.002 n=6)
TrackerStoreTrackSeries/totalSeries=100M/seriesPerRequest=10K/tenants=1000/cleanups=false-12     107.66µ ±  7%    75.79µ ±  2%   -29.60% (p=0.002 n=6)
geomean                                                                                           283.2µ          119.6µ         -57.78%
```